Add dependency on librt for Linux + OSX and AndroidUtil for Android.
This fixes the link errors introduced by change
I28b8239a88a74ff1951b0d9772ae941e7d0e50ae.

In addition:
* Cleaned up .gitignore to ignore the CommonTests unit test executable and
  removed superfluous .keystore files.
* Fixed undefined NAN constant referenced by b2Stat on Windows.
* Fixed undefined isnan function referenced by CommonTests unit test on
  Windows.
* Fixed warnings in Sparky Testbed test.

Tested:
Verified all targets build successfully for Android, Linux, OSX and Windows.

+++ liquidfun/Box2D/Testbed/Tests/Sparky.h
@@ -48,7 +48,7 @@ public:
 		m_pg = m_world->CreateParticleGroup(pd);
 
 		m_initialLifetime = m_remainingLifetime = lifetime;
-		m_halfLifetime = m_initialLifetime * 0.5;
+		m_halfLifetime = m_initialLifetime * 0.5f;
 
 		// Set particle initial velocity based on how far away it is from origin,
 		// exploding outwards.
@@ -93,7 +93,9 @@ public:
 			for (int i = bufferIndex; i < bufferIndex + m_pg->GetParticleCount(); i++)
 			{
 				b2ParticleColor &c = colors[i];
-				c.Set(m_origColor.r * coeff, m_origColor.g * coeff, m_origColor.b * coeff, m_origColor.a);
+				c.Set((int32)(m_origColor.r * coeff),
+				      (int32)(m_origColor.g * coeff),
+				      (int32)(m_origColor.b * coeff), m_origColor.a);
 			}
 		}
 	}
